Module Octez_smart_rollup_node_PsParisC.Arith_pvm

module Arith_proof_format : sig ... end

This module manifests the proof format used by the Arith PVM as defined by the Layer 1 implementation for it.

module Impl : Pvm_sig.S

include module type of struct include Impl end

type repo = Impl.repo

type tree = Impl.tree

module Ctxt_wrapper = Impl.Ctxt_wrapper

include Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.PVM.S with type hash = Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.State_hash.t and type context = ([ `Read | `Write ], repo) Tezos_layer2_store.Context_sigs.raw_index and type state = tree

val parse_boot_sector : string -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.PVM.boot_sector option

val pp_boot_sector : Tezos_protocol_environment_020_PsParisC.Format.formatter -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.PVM.boot_sector -> unit

type state = tree

val pp : state -> (Tezos_protocol_environment_020_PsParisC.Format.formatter -> unit -> unit) Tezos_protocol_environment_020_PsParisC.Lwt.t

type context = ([ `Read | `Write ], repo) Tezos_layer2_store.Context_sigs.raw_index

type hash = Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.State_hash.t

type proof = Impl.proof

val proof_encoding : proof Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val proof_start_state : proof -> hash

val proof_stop_state : proof -> hash

val state_hash : state -> hash Tezos_protocol_environment_020_PsParisC.Lwt.t

val initial_state : empty:state -> state Tezos_protocol_environment_020_PsParisC.Lwt.t

val install_boot_sector : state -> string -> state Tezos_protocol_environment_020_PsParisC.Lwt.t

val is_input_state : is_reveal_enabled: Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.is_reveal_enabled -> state -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.input_request Tezos_protocol_environment_020_PsParisC.Lwt.t

val set_input : Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.input -> state -> state Tezos_protocol_environment_020_PsParisC.Lwt.t

val eval : state -> state Tezos_protocol_environment_020_PsParisC.Lwt.t

val verify_proof : is_reveal_enabled: Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.is_reveal_enabled -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.input option -> proof -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.input_request Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val produce_proof : context -> is_reveal_enabled: Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.is_reveal_enabled -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.input option -> state -> proof Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

type output_proof = Impl.output_proof

val output_proof_encoding : output_proof Tezos_protocol_environment_020_PsParisC.Data_encoding.t

val output_of_output_proof : output_proof -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.output

val state_of_output_proof : output_proof -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.State_hash.t

val verify_output_proof : output_proof -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.output Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult Tezos_protocol_environment_020_PsParisC.Lwt.t

val produce_output_proof : context -> state -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.output -> (output_proof, Tezos_protocol_environment_020_PsParisC.Error_monad.error) Tezos_protocol_environment_020_PsParisC.Pervasives.result Tezos_protocol_environment_020_PsParisC.Lwt.t

val check_dissection : default_number_of_sections:int -> start_chunk: Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.Dissection_chunk.t -> stop_chunk: Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.Dissection_chunk.t -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Sc_rollup.Dissection_chunk.t list -> unit Tezos_protocol_environment_020_PsParisC.Error_monad.tzresult

val get_current_level : state -> Tezos_raw_protocol_020_PsParisC.Alpha_context.Raw_level.t option Tezos_protocol_environment_020_PsParisC.Lwt.t

module Internal_for_tests = Impl.Internal_for_tests

val kind : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.Kind.t

Kind of the PVM.

val get_tick : state -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.Tick.t Lwt.t

get_tick state gets the total tick counter for the given PVM state.

type status = Impl.status

PVM status

val get_status : is_reveal_enabled: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.is_reveal_enabled -> state -> status Lwt.t

get_status ~is_reveal_enabled state gives you the current execution status for the PVM.

val string_of_status : status -> string

string_of_status status returns a string representation of status.

val get_outbox : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Raw_level.t -> state -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.output list Lwt.t

get_outbox outbox_level state returns a list of outputs available in the outbox of state at a given outbox_level.

val eval_many : ?check_invalid_kernel:bool -> reveal_builtins:Tezos_scoru_wasm.Builtins.reveals -> write_debug:Tezos_scoru_wasm.Builtins.write_debug -> is_reveal_enabled: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.is_reveal_enabled -> ?stop_at_snapshot:bool -> max_steps:int64 -> state -> (state * int64) Lwt.t

eval_many ~max_steps s0 returns a state s1 resulting from the execution of up to ~max_steps steps of the rollup at state s0.

val new_dissection : default_number_of_sections:int -> start_chunk: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.Dissection_chunk.t -> our_stop_chunk: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.Dissection_chunk.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Sc_rollup.Tick.t list

module State = Impl.State

State storage for this PVM.

module Inspect_durable_state = Impl.Inspect_durable_state

Inspect durable state using a more specialised way of reading the PVM state. For example in WASM, it decodes the durable storage in the state before reading values.

module Unsafe_patches = Impl.Unsafe_patches

Expose unsafe state patching functions for manual intervention. At the moment this feature is only used to increase the maximum number of ticks of the WASM PVM in a non refutable setting.