Module Octez_smart_rollup_node_PsParisC.Dal_pages_request

Access DAL slots and pages content.

This module is a wrapper on top of Store.Dal_slot_pages module to access DAL slots and pages' data that have been previously fetched by the rollup node.

type Tezos_base.TzPervasives.error +=

| Dal_slot_not_found_in_store of Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Dal.Slot.Header.id

This error is returned when a slot, identified by its ID, is not found in the store.

val slot_pages : Octez_smart_rollup.Rollup_constants.dal_constants -> dal_activation_level: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Raw_level.t option -> inbox_level:int32 -> _ Octez_smart_rollup_node.Node_context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Dal.slot_id -> dal_attested_slots_validity_lag:int -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Dal.Page.content list option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the pages' content of the given slot ID's from the store.

The function returns Dal_slot_not_found_in_store if no entry is found in the store for the given ID (i.e. no page is registered with or without content).

If the returned value is Some pages, the slot whose ID is given is supposed to be confirmed and pages correspond to the pages of the slot. Otherwise None is returned.

The function relies on Store.Dal_slot_pages's invariants to guarantee that:

the pages are returned in increasing order w.r.t. their indexes in the slot;

the size of the list, in case it is not empty, is equal to the expected number of pages in a slot.

dal_attestation_lag is used to retrieve the correct entry in store.

val page_content : Octez_smart_rollup.Rollup_constants.dal_constants -> dal_activation_level: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Raw_level.t option -> inbox_level:int32 -> _ Octez_smart_rollup_node.Node_context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Dal.Page.t -> dal_attested_slots_validity_lag:int -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Dal.Page.content option Tezos_base.TzPervasives.tzresult Lwt.t

Retrieve the content of the page identified by the given ID from the store.

The function returns Dal_slot_not_found_in_store if no entry is found in the store for the given ID. It returns None in case the level of the requested page is out of bounds (e.g. in the future). Said otherwise, some content is only returned for confirmed pages (slots) for which the content has already been downloaded and saved to the store.

dal_attestation_lag is used to retrieve the correct entry in store.