Module Octez_smart_rollup_node_PsParisC.Inbox_event

val get_messages : Tezos_base.TzPervasives.Block_hash.t -> int32 -> int -> unit Lwt.t

get_messages hash level n emits the event that n messages are being fetched from the block of the given hash at the given level.