Module Octez_smart_rollup_node_PsParisC.RPC_directory

val block_directory : Octez_smart_rollup_node.Node_context.rw -> (unit * Octez_smart_rollup.Rollup_node_services.Arg.block_id) Tezos_rpc.Directory.t

The RPC directory, specific to blocks, for this rollup node.

val directory : Octez_smart_rollup_node.Node_context.rw -> unit Tezos_rpc.Directory.t

The full RPC directory for this rollup node, merging the top level directory and the block directory.